+++ R/create_venv.R
@@ -11,7 +11,7 @@ provision_venv = function() {
 
   ## If no python packages listed don't do anything
   if (length(pkgs) == 0L) return(invisible(NULL))
-
+  msg_start("Creating a venv...provision_venv()")
   venv_path = file.path(get_root_dir(), "notes", "venv")
   ## If a virtualenv hasn't already been made, make it
   if (!dir.exists(venv_path)) {
@@ -20,6 +20,7 @@ provision_venv = function() {
 
   ## Activate the virtualenv
   reticulate::use_virtualenv(venv_path)
+  Sys.setenv("RETICULATE_PYTHON" = file.path(venv_path, "/bin/python"))
 }
 
 #' @importFrom reticulate virtualenv_create
@@ -30,8 +31,6 @@ create_venv = function(pkgs) {
   ## By default reticulate creates virtualenvs in ~/.virtualenv
   ## To change where the virtualenv is created we have to use
   ## the WORKON_HOME environment variable
-  workon_home = Sys.getenv("WORKON_HOME")
-  on.exit(Sys.setenv(WORKON_HOME = workon_home))
   Sys.setenv(WORKON_HOME = file.path(get_root_dir(), "notes"))
 
   ## Create the virtualenv